requirements.txt
requests
requests_oauthlib
dogpile.cache
